requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
jinja2==3.1.4

# API clients
requests==2.32.3
python-dotenv==1.0.1

# AssemblyAI (STT)
assemblyai==0.33.0

# Google Gemini (LLM)
google-generativeai==0.7.2

# Optional (but useful)
pydantic==2.8.2
